Created readn_grams.py
Contains a method for reading in n_Grams from https://www.ngrams.info/download_coca.asp.
Method returns a 2D array of the n_grams, stored as a string. If run as main, script will print out the contents of the file to show how the contents of the array. The array is sorted by frequency.